requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.8.0